Log Compute progress using logarithmic distance
log10dPace is the movement of d,
expressed as log10(d/d0) where d0 is the previous d
(or on first iteration, d0=1.0 - maximum possible Norm2).

log10dRemaining is the remaining logarithmic distance,
expressed as log10(d/e); Compute() finishes when this drops below 0.

This is easier to quickly grasp