Make (*CSMatrix).Mmap() handle partially mmaped
Some operations on Mmap()-ed matrix replace specific major spans with
an in-memory slice, which results in a hybrid matrix with mapped and
unmmaped spans.  Make Mmap() deal with such cases by boundary-checking
slice data addresses (with the unsafe package) to see if any span lies
outside the memory-mapped region, and if so, remap the entire contents
into a fresh memory mapping, discarding the old memory mapping.